% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio.R
\name{ratio}
\alias{ratio}
\title{Takes the ratio of two numbers}
\usage{
ratio(x, y, digits)
}
\arguments{
\item{x}{A numeric representing the numerator of the ratio}

\item{y}{A numeric representing the denominator of the ratio}

\item{n}{An integer specifying the number of digits the result will be rounded to}
}
\value{
A scalar
}
\description{
Takes two numeric scalar arguments and returns their ratio rounded to a specified numner of digits
}
